+++ docs/app/layout.tsx
@@ -1,9 +1,7 @@
-import { siteConfig } from '@/lib/config';
-import { b2bTheme, coreTheme, ruiTheme, theme } from '@/theme';
+import { theme } from '@/theme';
 import { MarigoldProvider } from '@/ui';
-import { Suspense } from 'react';
+import { ReactNode, Suspense } from 'react';
 import { fontSans } from '@/theme/fonts';
-import { MarigoldThemeSwitch } from '@/ui/ThemeSwitch';
 import { Analytics } from './_components/Analytics';
 import { PortalContaier } from './_components/PortalContainer';
 import './globals.css';
@@ -25,17 +23,9 @@ export const metadata = {
   },
 };
 
-// Themes
-// ---------------
-const themes = {
-  b2b: b2bTheme,
-  core: coreTheme,
-  rui: ruiTheme,
-};
-
 // Layout
 // ---------------
-const Layout = ({ children }: { children: React.ReactNode }) => {
+const Layout = ({ children }: { children: ReactNode }) => {
   return (
     <html
       lang="en"
@@ -43,15 +33,10 @@ const Layout = ({ children }: { children: React.ReactNode }) => {
     >
       <body className={`${fontSans.className} min-h-screen`}>
         <Suspense>
-          <MarigoldThemeSwitch
-            themes={themes}
-            initial={siteConfig.defaultTheme}
-          >
-            <MarigoldProvider theme={theme} className="min-h-screen">
-              {children}
-            </MarigoldProvider>
-            <PortalContaier />
-          </MarigoldThemeSwitch>
+          <MarigoldProvider theme={theme} className="bg-bg-body min-h-screen">
+            {children}
+          </MarigoldProvider>
+          <PortalContaier />
         </Suspense>
         <Analytics />
       </body>