+++ docs/content/components/content/divider/divider-bold.demo.tsx
@@ -1,9 +1,0 @@
-import { Divider } from '@marigold/components';
-
-export default () => (
-  <>
-    <div>Up</div>
-    <Divider variant="bold" />
-    <div>Down</div>
-  </>
-);