+++ packages/components/src/DatePicker/DatePicker.stories.tsx
@@ -1,7 +1,7 @@
 import { CalendarDate, getLocalTimeZone, today } from '@internationalized/date';
-import { useState } from '@storybook/preview-api';
 import type { Meta, StoryObj } from '@storybook/react';
 import { DateValue } from 'react-aria-components';
+import { useState } from 'storybook/preview-api';
 import { I18nProvider } from '@react-aria/i18n';
 import { Stack } from '../Stack';
 import { DatePicker } from './DatePicker';