+++ packages/components/src/SelectList/SelectList.stories.tsx
@@ -5,9 +5,6 @@ import { SelectList } from './SelectList';
 
 const meta = {
   title: 'Components/SelectList',
-  parameters: {
-    disableLabelWidth: true,
-  },
   argTypes: {
     selectionMode: {
       control: {