+++ themes/theme-b2b/src/components/ComboBox.styles.ts
@@ -1,5 +1,0 @@
-import { ThemeComponent, cva } from '@marigold/system';
-
-export const ComboBox: ThemeComponent<'ComboBox'> = cva(
-  'absolute right-2 size-4 border-none bg-transparent p-0'
-);