+++ themes/theme-b2b/src/mixins.ts
@@ -1,5 +1,0 @@
-/**
- * Reusable set of styles (Tailwind classes).
- */
-
-export const ELEVALTION_RING = 'ring-1 ring-gray-600/5';