+++ themes/theme-rui/src/components/Divider.styles.ts
@@ -1,13 +1,13 @@
 import { ThemeComponent, cva } from '@marigold/system';
 
-export const Divider: ThemeComponent<'Divider'> = cva(
-  'bg-stone-300 h-px w-full',
-  {
-    variants: {
-      variant: {
-        bold: 'h-0.5',
-        section: '',
-      },
+export const Divider: ThemeComponent<'Divider'> = cva('bg-border h-px w-full', {
+  variants: {
+    variant: {
+      default: '',
+      bold: 'h-0.5',
     },
-  }
-);
+  },
+  defaultVariants: {
+    variant: 'default',
+  },
+});